requirements.txt
psycopg2==2.9.11
pydantic==1.10.8
PyYAML==6.0
superagi-tools==1.0.8
typing_extensions==4.15.0
